extract html template and js code
