+++ paper2html/templates/stylesheet.css
@@ -1,0 +1,30 @@
+html, body {
+    height: 100%;
+    overflow: hidden;
+    margin: 0;
+}
+#split{
+    height: 100%;
+}
+#left {
+    float: left;
+    top: 0;
+    width: 50%;
+    height: 100%;
+    overflow: auto;
+    box-sizing: border-box;
+    z-index: 1;
+    padding: 0%;
+}
+#right{
+    float: left;
+    top: 0;
+    left: 50%;
+    width: 50%;
+    height: 100%;
+    overflow: auto;
+    box-sizing: border-box;
+    z-index: 2;
+    background-color: #FFFFFF;
+    padding: 0% 1.5em;
+}
