+++ paper2html/templates/two_panes_with_patch.html
@@ -1,0 +1,30 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+    <meta http-equiv="Content-type" content="text/html;charset=utf-8" />
+    <link href="{}" rel="stylesheet" type="text/css" />
+    <title>
+        {}
+    </title>
+</head>
+<body>
+<div id="split">
+    <header style="position: fixed;">
+        <input type="button" value="30%" onclick="Zoom(0.33);"/>
+        <input type="button" value="50%" onclick="Zoom(0.5);"/>
+        <input type="button" value="65%" onclick="Zoom(0.65);"/>
+        <input type="button" value="100%" onclick="Zoom(1);"/>
+        <a href="{}">[Original PDF]</a>
+    </header><br />
+    <div id="left">
+        {}
+    </div>
+    <div id="right">
+        {}
+    </div>
+</div>
+<script language="javascript" type="text/javascript">
+    {}
+</script>
+</body>
+</html>
