+++ paper2html/templates/two_panes_with_zoom.html
@@ -1,0 +1,26 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+    <meta http-equiv="Content-type" content="text/html;charset=utf-8" />
+    <link href="{}" rel="stylesheet" type="text/css" />
+    <title>
+        {}
+    </title>
+</head>
+<body>
+<div id="split">
+    <header id="header" style="position: fixed;">
+        <a href="{}">[Original PDF]</a>
+    </header><br />
+    <div id="left">
+        <canvas id="canvas"></canvas>
+    </div>
+    <div id="right">
+        {}
+    </div>
+</div>
+<script language="javascript" type="text/javascript">
+    {}
+</script>
+</body>
+</html>
