+++ setup.py
@@ -34,6 +34,7 @@ setup(
     long_description=long_description,
     long_description_content_type="text/markdown",
     packages=find_packages(),
+    package_data={"paper2html": ["templates/*.html", "templates/*.css", "templates/*.js"]},
     install_requires=[
         'pdf2image',
         'pdfminer.six >= 20200726',
